Move legacy slotting functions out of main code into unit tests
