+++ lib/core/test/unit/Cardano/Wallet/Primitive/TypesSpec.hs
@@ -30,7 +30,7 @@ import Cardano.Wallet.Primitive.AddressDerivation
     ( Depth (..), WalletKey (..), digest, publicKey )
 import Cardano.Wallet.Primitive.AddressDerivation.Jormungandr
     ( JormungandrKey (..), generateKeyFromSeed )
-import Cardano.Wallet.Primitive.Slotting
+import Cardano.Wallet.Primitive.Slotting.Legacy
     ( SlotParameters (..)
     , epochStartTime
     , flatSlot
